requirements.txt
MetaTrader5==5.0.4424
numpy<2
python-dotenv==1.0.1
websocket-client==1.8.0
